Oprava kruhového importu
