+++ webclient/core/logging_filters.py
@@ -3,7 +3,7 @@ import logging
 
 class UserLogFilter(logging.Filter):
     def filter(self, record):
-        from core.middleware import LogMiddleware
+        from core.log_middleware import LogMiddleware
 
         record.url = LogMiddleware.get_request_url()
         record.user_id = LogMiddleware.get_user_id() or "Anonymous"
